- Fixed inline tasks. - Fixed size query for task head attachment shader data size. - Fixed shader integration sample. - Updated task declaration api to be like it was before in some parts.

+++ include/daxa/command_recorder.hpp
@@ -14,7 +14,7 @@ namespace daxa
     struct PushConstantInfo
     {
         void const * data = {};
-        u32 size = {};
+        u64 size = {};
         u32 offset = {};
     };
